extract styles into their own file
