+++ src/UIComponents/CrossTab.jsx
@@ -6,7 +6,7 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { getCrossTabData } from "../redux";
-import { styles } from "../constants.js";
+import { styles } from "../styles.js";
 import ScrollableContent from "./ScrollableContent";
 
 class CrossTab extends Component {
