+++ src/UIComponents/DataTable.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { setCurrentColumn, setHighlightColumn } from "../redux";
-import { styles } from "../constants";
+import { styles } from "../styles";
 
 class DataTable extends Component {
   static propTypes = {
