+++ src/UIComponents/ModelCard.jsx
@@ -2,7 +2,7 @@
 import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { styles } from "../constants";
+import { styles } from "../styles";
 import {
   getPercentCorrect,
   getLabelToSave,
