+++ src/UIComponents/Results.jsx
@@ -2,7 +2,7 @@
 import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { styles } from "../constants";
+import { styles } from "../styles";
 import { UnconnectedStatement } from "./Statement";
 import { setShowResultsDetails, setResultsPhase } from "../redux";
 import ResultsDetails from "./ResultsDetails";
