+++ src/UIComponents/ScatterPlot.jsx
@@ -2,7 +2,7 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { getScatterPlotData } from "../redux";
-import { styles, colors } from "../constants.js";
+import { styles, colors } from "../styles";
 import { Scatter } from "react-chartjs-2";
 
 const scatterDataBase = {
