+++ src/UIComponents/Statement.jsx
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import React, { Component } from "react";
 import { connect } from "react-redux";
 import { setLabelColumn, removeSelectedFeature } from "../redux";
-import { styles } from "../constants";
+import { styles } from "../styles";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faTimesCircle } from "@fortawesome/free-solid-svg-icons";
